Fix Crash Ratios report
Github issue 233

Update Crash Ratios to work with new data and frameworks

UI Changes:

- add version to PageState
- minor changes to controls and html ids

CLI Changes:

- added `update-versions-from-crashes` script, to be run daily

Migrations:

- remove `github_repo` and `github_issue_number` from `dtl.crashes`
- update `canonical_platforms` function to work with Brave Core versions
- add Versions table
- restore `dw.fc_crashes_dau_mv`, since updating `canonical_platforms`
dropped it

API & Model

- added `api/1/versions` endpoint to get consolidated list of versions
  - renamed old `versions` endpoint to `dau_by_version`
- updated `crash_ratios` endpoint and query to work with new data &
tables
- added versions model
- added crash version to canonical version mapping on Crash model

Tests:

- Feature/cucumber test for viewing data in Crash Ratios report
- api endpoint tests
  - versions endpoint
  - crash ratios endpoint
- model tests:
  - versions attributes
  - versions associations
  - crash version mapping functions

NPM:
  - added `vue` and `vue-loader` for webpack. I will use these in crash
  reporting

*Notes*

Several headaches with this one. One was the wild goose hunt through
databases when there was a test setup issue. The other was a selenium
issue. For some unknown reason, the front-end Application instance was
not receiving the `uiChange` event from the version select. For some
reason, this problem happens only when the browser is automated by
selenium.